ok I really understand it now.
Formatting in this framework (or also other frameworks, idfk) is handled in the global.css file. See the random jibberish that you can't find when searching for it (like <div class="wrapper stack gap-8 lg:gap-20">)?  Apparenlty each of those words is a sort of directive in the global.css. It confused me because they weren't seperated by commas.